+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 3.9.1
 - `BlockMovementCollision` improvements.
 - Create a `PinchGesture` mixin to listen pinch gestures.
-- Create a `UpdateCameraByPinchGesture` mixin to update zoom and camera position in pich events.
+- Create a `UpdateCameraByPinchGesture` mixin to update zoom and camera position in pinch events.
 
 # 3.9.0
 - `JoystickDirectional` improvements. Now you can use `Alignment`.