+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 3.9.2
 - MiniMap improviments. Fix issue [#517](https://github.com/RafaelBarbosatec/bonfire/issues/517)
 - Raname `BouncingObject` to `ElasticCollision`.